NMS-12383: add TimeseriesRoundTripIT Test, minor renamings
