+++ core/schema/src/main/liquibase/26.0.0/changelog.xml
@@ -88,4 +88,22 @@
                                  referencedTableName="graph_elements" referencedColumnNames="id" onDelete="CASCADE"/>
     </changeSet>
 
+    <changeSet author="pschweizer" id="26.0.0-create-timeseries_meta-table">
+        <createTable tableName="timeseries_meta">
+            <column name="resourceid" type="text">
+                <constraints nullable="false"/>
+            </column>
+            <column name="name" type="text">
+                <constraints nullable="false"/>
+            </column>
+            <column name="value" type="text">
+                <constraints nullable="false"/>
+            </column>
+        </createTable>
+        <addUniqueConstraint tableName="timeseries_meta" columnNames="resourceid,name" />
+
+        <rollback>
+            <dropTable tableName="timeseries_meta"/>
+        </rollback>
+    </changeSet>
 </databaseChangeLog>
