+++ features/timeseries/README.md
@@ -16,7 +16,8 @@
 * make sure the timescale plugin ist installed: ``select * from pg_extension;``
 
 ## TODOs
-* Understand difference between index and insert
+* disuss with Jesse the osgi compile problem in IntelliJ
+* finish TimeseriesRoundtripIT
 * Reading metadata from postgres
 * remove all Newts stuff from Abstraction layer
 * The code makes heavy use of the lombok library. This makes for much faster prototyping by remvoving lots of boilerplate code. Since we haven't agreed on using this library we might need to remove it (shouldn't be a problem)
